Add slate configs
Slate is a window management tool: https://github.com/jigish/slate

It is extremely configurable and its configs are stored in a .slate.js file in the users home directory.

Currently it automatically places application windows in correct locations for single and double monitor setups. It detects when monitor setups change and resets the window positions automatically.

It's hella cool.