Don't allow pending items in atom
Atom changed the behavior of opening files from the tree view. A file used to open on single click, but now that opens a temporary tab that will be replaced when single-clicking on a different file. Double-clicking the file, using the fuzzy finder, or editing the temporary file makes the file permanent.

This change removes the temporary file behavior. Single clicking opens nothing. This way, any file that is open is permanent, which I prefer.