Remove file extensions from aliases
I never end up using these in practice and occasionally they get in the way. For example, if I have a ruby file that I'm trying to run as a shell command, I would type something like: ./foo.rb but that would get interpreted as "open a ruby file" not "run this".