Add commit message hook to the default git template
This makes sure commit messages fold at 72 characters. When the commit
message is altered, it's piping into itself through the fold command. If
sponge is not used, it beings to write to itself, which erases the whole
file, then starts to read a now empty file. Sponge "soaks" up all of the
output first, allowing fold to do it's thing.