Switch path to be a function not alias
zsh aliases are expanded when their body is parsed, not when the alias is executed. That means if your alias prints out the contents of $PATH, anything added to the $PATH after setup won't be there. Instead I created a function that does the same thing but runs fresh every time.

Reference: http://zsh.sourceforge.net/Intro/intro_4.html